# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "jobtracker-api",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/source-groups": {
      "get": {
        "summary": "List Source Groups",
        "operationId": "list_source_groups_v1_source_groups_get",
        "parameters": [
          {
            "name": "x-service-token",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Service-Token"
            }
          },
          {
            "name": "x-user-sub",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Sub"
            }
          },
          {
            "name": "x-user-email",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Email"
            }
          },
          {
            "name": "x-user-name",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Name"
            }
          },
          {
            "name": "x-user-groups",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Groups"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/user/sources/apply-group": {
      "post": {
        "summary": "Apply Source Group",
        "operationId": "apply_source_group_v1_user_sources_apply_group_post",
        "parameters": [
          {
            "name": "x-service-token",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Service-Token"
            }
          },
          {
            "name": "x-user-sub",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Sub"
            }
          },
          {
            "name": "x-user-email",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Email"
            }
          },
          {
            "name": "x-user-name",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Name"
            }
          },
          {
            "name": "x-user-groups",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Groups"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyGroupBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/user/source-requests": {
      "post": {
        "summary": "Create Source Request",
        "operationId": "create_source_request_v1_user_source_requests_post",
        "parameters": [
          {
            "name": "x-service-token",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Service-Token"
            }
          },
          {
            "name": "x-user-sub",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Sub"
            }
          },
          {
            "name": "x-user-email",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Email"
            }
          },
          {
            "name": "x-user-name",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Name"
            }
          },
          {
            "name": "x-user-groups",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Groups"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceRequestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Own Source Requests",
        "operationId": "list_own_source_requests_v1_user_source_requests_get",
        "parameters": [
          {
            "name": "x-service-token",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Service-Token"
            }
          },
          {
            "name": "x-user-sub",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Sub"
            }
          },
          {
            "name": "x-user-email",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Email"
            }
          },
          {
            "name": "x-user-name",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Name"
            }
          },
          {
            "name": "x-user-groups",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Groups"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/sources": {
      "get": {
        "summary": "List Sources",
        "operationId": "list_sources_v1_sources_get",
        "parameters": [
          {
            "name": "x-service-token",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Service-Token"
            }
          },
          {
            "name": "x-user-sub",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Sub"
            }
          },
          {
            "name": "x-user-email",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Email"
            }
          },
          {
            "name": "x-user-name",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Name"
            }
          },
          {
            "name": "x-user-groups",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Groups"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/user/sources": {
      "put": {
        "summary": "Put Sources",
        "operationId": "put_sources_v1_user_sources_put",
        "parameters": [
          {
            "name": "x-service-token",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Service-Token"
            }
          },
          {
            "name": "x-user-sub",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Sub"
            }
          },
          {
            "name": "x-user-email",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Email"
            }
          },
          {
            "name": "x-user-name",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Name"
            }
          },
          {
            "name": "x-user-groups",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Groups"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourcesPut"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApplyGroupBody": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "default": "replace"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ApplyGroupBody"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SourceRequestBody": {
        "properties": {
          "url": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 8,
            "title": "Url"
          },
          "note": {
            "type": "string",
            "maxLength": 2000,
            "title": "Note",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "SourceRequestBody"
      },
      "SourcesPut": {
        "properties": {
          "enabled": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Enabled"
          }
        },
        "type": "object",
        "required": [
          "enabled"
        ],
        "title": "SourcesPut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}